+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/wait/wait_test.go
@@ -1,43 +1,59 @@
 package wait
 
 import (
-	"fmt"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/binding_constructors"
-	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/facts_engine"
+	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/service_network"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/kurtosis_instruction"
 	"github.com/stretchr/testify/require"
 	"go.starlark.net/starlark"
+	"go.starlark.net/starlarkstruct"
 	"testing"
 )
 
 const (
-	testServiceId = "example-service-id"
-	testFactName  = "example-fact-name"
+	testTargetKey   = "key"
+	testAssertion   = "=="
+	testTargetValue = "value"
+	testDurationStr = "5s"
+	testUuid        = "88a40d8d-8683-439e-ae16-45ea58b635ae"
 )
 
 var (
-	emptyFactsEngine *facts_engine.FactsEngine = nil
+	testTarget          = starlark.MakeInt(0)
+	emptyServiceNetwork = service_network.NewEmptyMockServiceNetwork()
 )
 
-func TestWaitInstruction_GetCanonicalizedInstruction(t *testing.T) {
-	position := kurtosis_instruction.NewInstructionPosition(1, 1, "dummyFile")
-	waitInstruction := newEmptyWaitInstruction(emptyFactsEngine, position)
-	waitInstruction.starlarkKwargs = starlark.StringDict{}
-	waitInstruction.starlarkKwargs[serviceIdArgName] = starlark.String(testServiceId)
-	waitInstruction.starlarkKwargs[factNameArgName] = starlark.String(testFactName)
-
-	expectedFormatStr := `wait(fact_name="%v", service_id="%v")`
-	expectedStr := fmt.Sprintf(expectedFormatStr, testFactName, testServiceId)
-	require.Equal(t, expectedStr, waitInstruction.String())
-
-	canonicalInstruction := binding_constructors.NewStarlarkInstruction(
-		position.ToAPIType(),
-		WaitBuiltinName,
-		expectedStr,
-		[]*kurtosis_core_rpc_api_bindings.StarlarkInstructionArg{
-			binding_constructors.NewStarlarkInstructionKwarg(`"`+string(testServiceId)+`"`, serviceIdArgName, true),
-			binding_constructors.NewStarlarkInstructionKwarg(`"`+testFactName+`"`, factNameArgName, true),
-		})
-	require.Equal(t, canonicalInstruction, waitInstruction.GetCanonicalInstruction())
+func TestWaitInstruction_StringRepresentationWorks(t *testing.T) {
+	extractor := &starlark.Dict{}
+	err := extractor.SetKey(starlark.String(testTargetKey), starlark.String(".value"))
+	require.Nil(t, err)
+	testRecipeConfig := starlarkstruct.FromStringDict(starlarkstruct.Default, starlark.StringDict{
+		"service_id":   starlark.String("web-server"),
+		"port_id":      starlark.String("http-port"),
+		"endpoint":     starlark.String("/"),
+		"method":       starlark.String("POST"),
+		"content_type": starlark.String("text/plain"),
+		"body":         starlark.String("post_output"),
+		"extract":      extractor,
+	})
+	starlarkKwargs := starlark.StringDict{
+		"recipe":       testRecipeConfig,
+		"assertion":    starlark.String(testAssertion),
+		"target_value": testTarget,
+		"timeout?":     starlark.String(testDurationStr),
+		"interval?":    starlark.String(testDurationStr),
+	}
+	starlarkKwargs.Freeze()
+	getValueInstruction := newWaitInstructionInstructionForTest(
+		emptyServiceNetwork,
+		kurtosis_instruction.NewInstructionPosition(1, 1, "dummyFile"),
+		nil,
+		nil,
+		testUuid,
+		testTargetKey,
+		testAssertion,
+		starlark.String(testTargetValue),
+		starlarkKwargs,
+	)
+	expectedStr := `wait(assertion="==", interval?="5s", recipe=struct(body="post_output", content_type="text/plain", endpoint="/", extract={"key": ".value"}, method="POST", port_id="http-port", service_id="web-server"), target_value=0, timeout?="5s")`
+	require.Equal(t, expectedStr, getValueInstruction.String())
 }