+++ core/server/api_container/server/startosis_engine/startosis_interpreter_test.go
@@ -67,6 +67,28 @@ print("` + testString + `")
 	validateScriptOutputFromPrintInstructions(t, instructions, expectedOutput)
 }
 
+func TestStartosisInterpreter_DefineFactAndWait(t *testing.T) {
+	packageContentProvider := mock_package_content_provider.NewMockPackageContentProvider()
+	defer packageContentProvider.RemoveAll()
+	interpreter := NewStartosisInterpreterWithFacts(testServiceNetwork, nil, packageContentProvider, runtime_value_store.NewRuntimeValueStore())
+	script := `
+get_recipe = struct(
+	service_id = "web-server",
+	port_id = "http-port",
+	endpoint = "?input=output",
+	method = "GET",
+	extract = {
+		"input": ".query.input"
+	}
+)
+response = wait(get_recipe, "code", "==", 200, timeout="5m", interval="5s")
+print(response["body"])
+`
+	_, instructions, interpretationError := interpreter.Interpret(context.Background(), startosis_constants.PackageIdPlaceholderForStandaloneScript, script, startosis_constants.EmptyInputArgs)
+	require.Nil(t, interpretationError)
+	require.NotEmpty(t, instructions)
+}
+
 func TestStartosisInterpreter_ScriptFailingSingleError(t *testing.T) {
 	packageContentProvider := mock_package_content_provider.NewMockPackageContentProvider()
 	defer packageContentProvider.RemoveAll()
@@ -914,23 +936,6 @@ print(file_contents)
 this is a test string
 `
 	validateScriptOutputFromPrintInstructions(t, instructions, expectedOutput)
-}
-
-func TestStartosisInterpreter_DefineFactAndWait(t *testing.T) {
-	packageContentProvider := mock_package_content_provider.NewMockPackageContentProvider()
-	defer packageContentProvider.RemoveAll()
-	interpreter := NewStartosisInterpreterWithFacts(testServiceNetwork, nil, packageContentProvider, nil)
-	scriptFormatStr := `
-define_fact(service_id="%v", fact_name="%v", fact_recipe=struct(method="GET", endpoint="/", port_id="http"))
-wait(service_id="%v", fact_name="%v")
-`
-	serviceId := "service"
-	factName := "fact"
-	script := fmt.Sprintf(scriptFormatStr, serviceId, factName, serviceId, factName)
-	_, instructions, interpretationError := interpreter.Interpret(context.Background(), startosis_constants.PackageIdPlaceholderForStandaloneScript, script, startosis_constants.EmptyInputArgs)
-	require.Nil(t, interpretationError)
-	require.NotEmpty(t, instructions)
-	validateScriptOutputFromPrintInstructions(t, instructions, "")
 }
 
 func TestStartosisInterpreter_RenderTemplates(t *testing.T) {