+++ internal_testsuites/golang/testsuite/startosis_request_wait_assert_test/startosis_request_wait_assert.go
@@ -9,7 +9,7 @@ import (
 )
 
 const (
-	testName              = "startosis_recipe_get_value_test"
+	testName              = "startosis_request_wait_assert_test"
 	isPartitioningEnabled = false
 	defaultDryRun         = false
 
@@ -33,7 +33,7 @@ def run(args):
 			"exploded-slash": ".query.input | split(\"/\") | .[1]"
 		}
 	)
-	response = request(get_recipe)
+	response = wait(get_recipe, "code", "==", 200, interval="10s", timeout="200s")
 	assert(response["code"], "==", 200)
 	assert("My test returned " + response["code"], "==", "My test returned 200")
 	assert(response["code"], "!=", 500)
@@ -55,6 +55,7 @@ def run(args):
 			"my-body": ".body"
 		}
 	)
+	wait(post_recipe, "code", "==", 200)
 	post_response = request(post_recipe)
 	assert(post_response["code"], "==", 200)
 	assert(post_response["extract.my-body"], "==", "post_output")